Splitting CLI related functionality off GUI classes ResourceManager and ConfigFragment.
